+++ sdk/keyvault/keyvault-secrets/src/index.ts
@@ -41,7 +41,7 @@ import { DeleteSecretPollOperationState } from "./lro/delete/operation";
 import { RecoverDeletedSecretPollOperationState } from "./lro/recover/operation";
 
 import {
-  Secret,
+  KeyVaultSecret,
   SecretClientInterface,
   SecretPollerOptions,
   DeletedSecret,
@@ -72,7 +72,7 @@ export {
   ParsedKeyVaultEntityIdentifier,
   PollerLike,
   PollOperationState,
-  Secret,
+  KeyVaultSecret,
   SecretProperties,
   SecretPollerOptions,
   SetSecretOptions,
@@ -198,7 +198,7 @@ export class SecretClient {
     secretName: string,
     value: string,
     options?: SetSecretOptions
-  ): Promise<Secret> {
+  ): Promise<KeyVaultSecret> {
     if (options) {
       const unflattenedProperties = {
         enabled: options.enabled,
@@ -298,7 +298,7 @@ export class SecretClient {
     secretName: string,
     secretVersion: string,
     options?: UpdateSecretOptions
-  ): Promise<Secret> {
+  ): Promise<KeyVaultSecret> {
     if (options) {
       const unflattenedProperties = {
         enabled: options.enabled,
@@ -355,7 +355,7 @@ export class SecretClient {
    * @param secretName The name of the secret.
    * @param [options] The optional parameters
    */
-  public async getSecret(secretName: string, options?: GetSecretOptions): Promise<Secret> {
+  public async getSecret(secretName: string, options?: GetSecretOptions): Promise<KeyVaultSecret> {
     const span = this.createSpan("getSecret", options && options.requestOptions);
     const requestOptions = this.setParentSpan(span, options && options.requestOptions);
 
@@ -533,7 +533,7 @@ export class SecretClient {
   public async restoreSecret(
     secretBundleBackup: Uint8Array,
     options?: RequestOptionsBase
-  ): Promise<Secret> {
+  ): Promise<KeyVaultSecret> {
     const span = this.createSpan("restoreSecret", options);
 
     let response: RestoreSecretResponse;
@@ -852,7 +852,7 @@ export class SecretClient {
     };
   }
 
-  private getSecretFromSecretBundle(secretBundle: SecretBundle | DeletedSecretBundle): Secret {
+  private getSecretFromSecretBundle(secretBundle: SecretBundle | DeletedSecretBundle): KeyVaultSecret {
     const parsedId = parseKeyvaultEntityIdentifier("secrets", secretBundle.id);
 
     let resultObject;